+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Update `mio` dependency
 - [server] Fix some leaks when dropping the display when using the rust implementation
 
 ## 0.21.7 -- 2018-11-28